SQL_Query_Generator/requirements.txt
streamlit 
pandas
faiss-cpu
numpy
sentence_transformers
transformers